ai_engine/requirements.txt
transformers
torch
torchvision
pillow
ftfy
regex
git+https://github.com/openai/CLIP.git
